+++ wrappers/src/versioned_multi_key/mapx_triple_key/mod.rs
@@ -48,7 +48,7 @@ where
     #[inline(always)]
     pub fn new() -> Self {
         MapxTkVs {
-            inner: MapxRawMkVs::new(KEY_SIZE),
+            inner: MapxRawMkVs::new(KEY_SIZE as u32),
             p: PhantomData,
         }
     }
